Add finishing cleaning up imports for chartHelpers relocation
